requirements.txt
# LangChain 核心
langchain>=0.1.0
langchain-community>=0.0.10
langchain-openai>=0.0.5

# 向量数据库
chromadb>=0.4.0

# 嵌入模型
sentence-transformers>=2.2.0

# 环境变量
python-dotenv>=1.0.0

# 其他工具
tiktoken>=0.5.0
